Update the webgl code and add some helper functions
