+++ .eslintrc
@@ -39,7 +39,7 @@
     "no-extra-bind": "error",
     "no-loop-func": "error",
     "no-magic-numbers": ["warn", {
-      "ignore": [0, 1, -1],
+      "ignore": [0, 1, -1, 2],
       "ignoreArrayIndexes": true
     }],
     "no-multiple-empty-lines": ["warn", {
